image: Fix internal type of length parameter
This is a follow up change to 51d107218522 ("image: Define
PY_SSIZE_T_CLEAN and fix length types").

+++ sdaps/image/surface.c
@@ -152,7 +152,7 @@ set_pixels_unchecked(guint32* pixels, guint32 stride, gint x, gint y, gint width
 }
 
 void
-get_pbm(cairo_surface_t *surface, void **data, int *length)
+get_pbm(cairo_surface_t *surface, void **data, gssize *length)
 {
 	int width, height;
 	int s_stride;
